requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
#awscli
flake8
python-dotenv>=0.5.1
dvc
dvc_gdrive
cookiecutter
numpy
pandas
matplotlib
seaborn
scikit-learn==1.2.2
joblib
pyyaml
ipykernel
pytest
tox
flask
gunicorn
watchdog